refactor(snapshot): collapse workload into protocol
