--- deploy/snapshot/protocol/storage.go
@@ -1,57 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package protocol
-
-import (
-	"fmt"
-	"strings"
-)
-
-type Storage struct {
-	Type     string
-	Location string
-	PVCName  string
-	BasePath string
-}
-
-func ResolveCheckpointStorage(checkpointID string, version string, storage Storage) (Storage, error) {
-	resolved, err := resolveStorageConfig(storage)
-	if err != nil {
-		return Storage{}, err
-	}
-	resolved.Location = strings.TrimRight(resolved.BasePath, "/") + "/" + checkpointID + "/versions/" + NormalizeArtifactVersion(version)
-	return resolved, nil
-}
-
-func ResolveRestoreStorage(checkpointID string, version string, location string, storage Storage) (Storage, error) {
-	resolved, err := resolveStorageConfig(storage)
-	if err != nil {
-		return Storage{}, err
-	}
-	location = strings.TrimSpace(location)
-	if location == "" {
-		return ResolveCheckpointStorage(checkpointID, version, storage)
-	}
-	resolved.Location = location
-	return resolved, nil
-}
-
-func resolveStorageConfig(storage Storage) (Storage, error) {
-	storageType := strings.TrimSpace(storage.Type)
-	if storageType == "" {
-		storageType = StorageTypePVC
-	}
-	if storageType != StorageTypePVC {
-		return Storage{}, fmt.Errorf("checkpoint storage type %q is not supported", storageType)
-	}
-	basePath := strings.TrimSpace(storage.BasePath)
-	if basePath == "" {
-		return Storage{}, fmt.Errorf("checkpoint base path is required")
-	}
-	return Storage{
-		Type:     storageType,
-		PVCName:  strings.TrimSpace(storage.PVCName),
-		BasePath: strings.TrimRight(basePath, "/"),
-	}, nil
-}
